Adds byte output from websocket frame and README on decoding.
